refactor: Collect all Default I/O into `io.default`
This taskes the parts of the code implementing the Default I/O class,
which have been scattered around `alpenhorn.io` and collects them into
all into a `io.default` submodule to better organize the code.

The former `io._default_asyncs` and `io.ioutil` have been refactored
into operation-focused `io.default.{check,delete,pull}` modules.

The high-level Default I/O class definition have been dispersed into
`io.default.{group,node,remote}`.  The `default` module aslo now
contains the directory-tree locking mechanism `updownlock`.

Two bits of the old `ioutil` which weren't actually doing any I/O but
were concerned with database updates after a pull (mosly independent
of what sort of I/O class did the pull) have been moved to
`daemon.pullutil`.

I think this does a much better job of organising the `io` module.
Tests have been re-arranged accordingly.

Also, separately, I've renamed `io.lfs` to `io._lfs` to emphasise it's
not an I/O class, but a utility module used by the LustreQuota and
LustreHSM modules.
