+++ tests/io/default/test_node.py
@@ -284,7 +284,7 @@ def test_idle_cleanup_rate(unode, queue):
     """Test rate limiting of the cleanup task."""
 
     # This is how often the task should be queued
-    from alpenhorn.io.default import _IDLE_CLEANUP_PERIOD
+    from alpenhorn.io.default.node import _IDLE_CLEANUP_PERIOD
 
     # Run the idle update a few times with newly_idle true
     # The task is queued the first time, and then once
